agno-agent/requirements.txt
agno>=1.7.7
openai>=1.50.0
anthropic>=0.34.0
tavily-python>=0.5.0
duckduckgo-search>=6.0.0
yfinance>=0.2.0
python-dotenv>=1.0.0